journaling apis and tests
